requirements_latest.txt
opencv-contrib-python
opencv-python
numpy
Pillow
matplotlib
